Add Prim Path Picker + refactor some content around for Prim Hierarchy re-use
